src/mock-service/requirements.txt
fastapi==0.109.0
numpy==1.26.3
psycopg2==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.6
scipy==1.12.0
SQLAlchemy==2.0.25
uvicorn[standard]==0.27.0
gunicorn==21.2.0
